mapstruct

+++ ruoyi-poor-village/src/main/java/com/ruoyi/poor/dto/VillageDto.java
@@ -45,5 +45,5 @@ public class VillageDto {
     /**
      * 企业数量
      */
-    private Integer industryCount;
+    private Long industryCount;
 }
